Move `AccessibilityDetector` and `UpdateManager` to `com.osudroid` package
